Add plugin for LayerTree to show style switcher UI
This adds a plugin 'cmv_tree_column_style_switcher' to the LayerTree in
order to add a style switcher UI (radio buttons). The radio buttons are
derived from a styles config of the layer (e.g. 'sldStyles' for WMS).
With this commit only WMS are supported for a style switcher.