Add WFS support in cmv_tree_column_style_switcher
This enables the support of WFS layers to be used in the LayerTree
plugin 'cmv_tree_column_style_switcher'. The SLD files defined in the
config property 'styles' are listed as radio buttons and will be loaded
and applied to the layer in case the user checks a different radio.