Add Feature Selecton Tool
This adds a tool to select features on the map (by clicking) in order
to create an ExtJS filter having the clicked features in an "IN" filter
structure.
Allows to have a new selection at every click (resetting the internal
FID list) or adding a newly clicked feature to the filter.