Make style switcher usable in tree context menu
This makes the RadioGroup to switch between styles (normally rendered
under the corresponding layer tree node) usable in the context menu of
the tree node.
By setting the config property 'styleSwitcherBelowNode' to false for the
LayerTree will force the rendering in the context menu. Setting it to
true (default) will enforce the old behaviour.